aegis/requirements.txt
# Requirements for the Athena DRF gateway.  These dependencies form the
# foundation of a Django REST Framework project with real‑time
# capabilities via Django Channels and background task support via Celery.

Django==5.2.5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
channels>=4.0
channels-redis>=4.0
daphne>=4.1
celery>=5.3
kombu>=5.3
redis>=5.0
psycopg[binary,pool]==3.2.3
python-dotenv>=1.0
firebase-admin==6.5.0
python-telegram-bot>=21.0
## Hot-reload helpers for dev
watchfiles>=0.21
watchdog>=3.0

langchain
langchain-core
langchain-openai
langchain-postgres
langchain-mcp-adapters
langchain-community
langgraph
langgraph-prebuilt
langgraph-checkpoint-redis
langgraph-checkpoint-postgres
mcp
pgvector
sentence-transformers
langchain-cohere
cohere